Update Pong package
