+++ pong/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@darrenkuro/pong-core",
-    "version": "1.1.8",
+    "version": "1.1.9",
     "description": "Shared types and Pong engine for ft-transcendence project",
     "type": "module",
     "exports": {
