+++ pong/src/index.ts
@@ -1,9 +1,10 @@
-export * from "./types/types.ws";
+export * from "./schemas/types.ws";
 export * from "./pong/pong.engine";
 export * from "./pong/pong.types";
 export * from "./pong/pong.config";
 export * from "./utils";
-export * from "./types/schemas.api";
-export * from "./types/schemas.user";
+export * from "./schemas/schemas.api";
+export * from "./schemas/schemas.user";
+export * from "./schemas/schemas.totp";
 export * from "./ai/ai.service";
 export * from "./ai/ai.types";
